requirements.txt
behave==1.2.6
selenium==4.0.0
PyHamcrest>=2.0.2
webdriver-manager==3.2.2
requests~=2.32.3